CachingLinkGenerator & LinkGenerator: moved duplicated asset(Map) code to AssetProcessorService#asset(Map, DefaultLinkGenerator) in asset(Map), call assetProcessorService.asset(attrs, this)
